Use forked openid_connect module to decouple console from plural oidc
The oidc connect genserver will currently terminate if it cannot query the oidc config document, which can bring down console instances and couple our console to oour own availability.  In reality this is inevitable because people won't be able to log back in but we can at least presere existing sessions this way.