Console leader election
Currently we allow git updates to occur on any node in the console's cluster.  This can easily split brain and is not ideal, but needed to devise a specific leader election system to improve it. This uses postgres as coordinator, and allows for electing deployer pids as cluster leaders using a simple mechanism:

* any deployer requests leadership, first wins
* leader emits a heartbeat that expires after three missed beats
* on deployer teardown, force delete leader records to speed up reelection
* follower nodes rpc to the leader using distributed erlang, with fallback to themselves if none exists

Also added some more git cleaning logic