Implement the ability to trigger custom stack runs
This can be really useful in case you need to do one-off tasks like repair terraform state or run a bash script.  The run still will occur in the exec environment of the stack, ensuring consistency w/ the terraform run itself.

Also adds a CustomStackRun schema which adds a self-service layer on top of this.