Add caching for change detection w/ stacks
Stacks need to query git for changes each poll interval to determine if 1. there's a new sha and 2. if there's a file that it's tracking that changed in that new sha.  This can theoretically lead to a lot of contention in the git cache.  this has already been partially solved by adding a tracking column to determine if a sha has already been polled, this change cache will also solve for many stacks polling against the same folder and thus reduce duplicate work there as well.