Implement manifest digest api
This will allow the agent to determine if downloading manifests can be bypassed, hopefully giving another meaningful increase in overall throughput for the agent.  It is annoyingly complex due to the fact we source manifests from many locations, but it should mirror the manifest tar logic and incorporate all information into a consistently compiled sha.