Support In-App Notifications and Shared Secrets
This is a paired pr because they feed off each other, in app notifications give a simple api to support a notif inbox in-app.  We can then use that to provide a secret sharing experience, where you can share w/ a user/group list and give a one-time url in the inbox to read the secret.