Fix router filter logic
There was a slight misconfiguration in the left joining that was too clever for its own good.  Fix it to be more explicit about the required semantics, ie a notification router is eligible if either:

* it has no filters at all
* it has at least one router filter matching the criteria of the event

this is done by doing two joins and unioning them, which should be relatively efficient.