Implement sentinel integration job runner
This is likely the most powerful workflow we can tie into a sentinel, in effect its an integration test job that can be spawned across multiple clusters.  We'll need to write a decent amount of agent code to work alongside it, but this provides a lot of the api-side support.